feat : modif of structure in about component #1
