+++ client/src/components/About/dataAbout.ts
@@ -1,0 +1,51 @@
+import { Clock, Target, Trophy, Users } from "lucide-react";
+
+export const whoWeAre = [
+  {
+    icon: Users,
+    title: "Expertise IT",
+    description:
+      "En tant qu’expert du recrutement informatique, Externatic est là pour mieux vous connaître et pour vous guider dans vos choix.",
+  },
+  {
+    icon: Target,
+    title: "Accompagnement",
+    description:
+      "Vous accompagner, vous conseiller et vous aiguiller. Tel est notre métier.",
+  },
+  {
+    icon: Clock,
+    title: "Réactivité",
+    description:
+      "Les connexions que nous tissons, nous les entretenons sur la durée.",
+  },
+  {
+    icon: Trophy,
+    title: "Résultats",
+    description:
+      "Nous sommes acteurs de l’écosystème tech local, et nous nous appuyons sur un réseau riche d’expériences et d’expertises.",
+  },
+];
+
+export const services = [
+  {
+    title: "Au succès",
+    description:
+      "Nous fonctionnons “sans engagement”. Aucun candidat = pas d’acompte à verser",
+  },
+  {
+    title: "Aucune exclusivité",
+    description:
+      "Priorité à la satisfaction client : pleine liberté dans vos décisions d’accompagnement & partenariat.",
+  },
+  {
+    title: "Garantie 1ʳᵉ période d'essai",
+    description:
+      "Nous nous engageons durant cette période à retrouver un profil similaire à votre besoin sans frais supplémentaire.",
+  },
+  {
+    title: "Cherchons LE bon candidat",
+    description:
+      "Notre force réside dans notre réseau local, soutenu par nos engagements et cooptations.",
+  },
+];
